+++ packages/templates/src/api/index.ts
@@ -4,12 +4,12 @@ import eslintConfig from "./templates/eslint-config";
 import nestCli from "./templates/nest-cli";
 import tsconfig from "./templates/tsconfig";
 import tsconfigBuild from "./templates/tsconfig-build";
-import appController from "./templates/app-controller";
 import appModule from "./templates/app-module";
-import main from "./templates/main";
-import todosController from "./templates/todos-controller";
+import appService from "./templates/app-service";
+import appServiceSpec from "./templates/app-service-spec";
 import todosModule from "./templates/todos-module";
 import todosService from "./templates/todos-service";
+import todosServiceSpec from "./templates/todos-service-spec";
 import readme from "./templates/readme";
 
 const api: Module = {
@@ -19,12 +19,12 @@ const api: Module = {
     nestCli,
     tsconfig,
     tsconfigBuild,
-    appController,
     appModule,
-    main,
-    todosController,
+    appService,
+    appServiceSpec,
     todosModule,
     todosService,
+    todosServiceSpec,
     readme,
   ],
 };