+++ packages/templates/src/api/templates/todos-module.ts
@@ -1,12 +1,11 @@
 export default {
   filename: "apps/api/src/todos/todos.module.ts",
   template: `import { Module } from '@nestjs/common';
-import { TodosController } from './todos.controller';
 import { TodosService } from './todos.service';
 
 @Module({
-  controllers: [TodosController],
   providers: [TodosService],
+  exports: [TodosService],
 })
 export class TodosModule {}`,
 };