+++ packages/templates/src/auth/templates/package-json.ts
@@ -5,13 +5,9 @@ export default {
   "version": "0.0.0",
   "private": true,
   "exports": {
-    ".": {
-      "default": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    }
+    ".": "./src/index.ts"
   },
   "scripts": {
-    "build": "tsc",
     "migrate": "dotenv -e .env -- auth migrate --config src/index.ts",
     "generate": "dotenv -e .env -- auth generate --config src/index.ts"
   },