+++ packages/components/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @newt-app/components
 
+## 0.13.2
+
+### Patch Changes
+
+- @newt-app/ui@0.13.2
+
+## 0.13.1
+
+### Patch Changes
+
+- @newt-app/ui@0.13.1
+
+## 0.13.0
+
+### Patch Changes
+
+- @newt-app/ui@0.13.0
+
+## 0.12.0
+
+### Patch Changes
+
+- @newt-app/ui@0.12.0
+
+## 0.11.0
+
+### Patch Changes
+
+- @newt-app/ui@0.11.0
+
+## 0.10.0
+
+### Patch Changes
+
+- @newt-app/ui@0.10.0
+
+## 0.9.0
+
+### Patch Changes
+
+- @newt-app/ui@0.9.0
+
+## 0.8.3
+
+### Patch Changes
+
+- @newt-app/ui@0.8.3
+
+## 0.8.2
+
+### Patch Changes
+
+- @newt-app/ui@0.8.2
+
+## 0.8.1
+
+### Patch Changes
+
+- @newt-app/ui@0.8.1
+
+## 0.8.0
+
+### Patch Changes
+
+- @newt-app/ui@0.8.0
+
+## 0.7.0
+
+### Patch Changes
+
+- @newt-app/ui@0.7.0
+
 ## 0.6.0
 
 ### Patch Changes